Added success and error messages to character controller actions

+++ Trident/Controllers/CharacterController.cs
@@ -33,17 +33,28 @@ namespace Trident.Controllers
         [HttpPost]
         public ActionResult Create(string CharacterName_New, int CharacterWeapon_New, int CharacterTreasure_New, int? CharacterMember_New)
         {
-            //Query string 
-            string query = "insert into characters(CharacterName, CharacterWeapon, CharacterTreasure, member_MemberID) values (@name, @weapon, @treasure, @mid)";
+            if(ModelState.IsValid)
+            {
+                //Query string 
+                string query = "insert into characters(CharacterName, CharacterWeapon, CharacterTreasure, member_MemberID) values (@name, @weapon, @treasure, @mid)";
 
-            MySqlParameter[] myParams = new MySqlParameter[4];
-            myParams[0] = new MySqlParameter("@name", CharacterName_New);
-            myParams[1] = new MySqlParameter("@weapon", CharacterWeapon_New);
-            myParams[2] = new MySqlParameter("@treasure", CharacterTreasure_New);
-            myParams[3] = new MySqlParameter("@mid", CharacterMember_New);
+                MySqlParameter[] myParams = new MySqlParameter[4];
+                myParams[0] = new MySqlParameter("@name", CharacterName_New);
+                myParams[1] = new MySqlParameter("@weapon", CharacterWeapon_New);
+                myParams[2] = new MySqlParameter("@treasure", CharacterTreasure_New);
+                myParams[3] = new MySqlParameter("@mid", CharacterMember_New);
 
-            db.Database.ExecuteSqlCommand(query, myParams);
-            return RedirectToAction("Show/" + CharacterMember_New, "Member");
+                db.Database.ExecuteSqlCommand(query, myParams);
+
+                TempData["AddSuccess"] = "Character successfully added";
+                return RedirectToAction("Show/" + CharacterMember_New, "Member");
+
+            }
+            else
+            {
+                TempData["AddFail"] = "Failed to add character";
+                return RedirectToAction("Show/" + CharacterMember_New, "Member");
+            }
         }
 
         [Authorize(Roles = "Member, Admin")]
@@ -80,52 +91,71 @@ namespace Trident.Controllers
         [HttpPost]
         public ActionResult Edit(int? id, string CharacterName, int CharacterWeapon, int CharacterTreasure, int? CharacterMember)
         {
-            if ((id == null) || (db.Characters.Find(id) == null))
+            if(ModelState.IsValid)
             {
-                return HttpNotFound();
+                if ((id == null) || (db.Characters.Find(id) == null))
+                {
+                    return HttpNotFound();
+                }
+
+                string query = "update characters set CharacterName=@name, " +
+                    "CharacterWeapon=@weapon,"+
+                    "CharacterTreasure=@treasure," +
+                    "member_MemberID=@mid where CharacterID=@id";
+
+                MySqlParameter[] myParams = new MySqlParameter[5];
+                myParams[0] = new MySqlParameter();
+                myParams[0].ParameterName = "@name";
+                myParams[0].Value = CharacterName;
+
+                myParams[1] = new MySqlParameter();
+                myParams[1].ParameterName = "@weapon";
+                myParams[1].Value = CharacterWeapon;
+
+                myParams[2] = new MySqlParameter();
+                myParams[2].ParameterName = "@treasure";
+                myParams[2].Value = CharacterTreasure;
+
+                myParams[3] = new MySqlParameter();
+                myParams[3].ParameterName = "mid";
+                myParams[3].Value = CharacterMember;
+
+                myParams[4] = new MySqlParameter();
+                myParams[4].ParameterName = "@id";
+                myParams[4].Value = id;
+
+                db.Database.ExecuteSqlCommand(query, myParams);
+
+                TempData["EditSuccess"] = "Character successfully edited";
+                return RedirectToAction("Show/" + id);
+                
             }
-
-            string query = "update characters set CharacterName=@name, " +
-                "CharacterWeapon=@weapon,"+
-                "CharacterTreasure=@treasure," +
-                "member_MemberID=@mid where CharacterID=@id";
-
-            MySqlParameter[] myParams = new MySqlParameter[5];
-            myParams[0] = new MySqlParameter();
-            myParams[0].ParameterName = "@name";
-            myParams[0].Value = CharacterName;
-
-            myParams[1] = new MySqlParameter();
-            myParams[1].ParameterName = "@weapon";
-            myParams[1].Value = CharacterWeapon;
-
-            myParams[2] = new MySqlParameter();
-            myParams[2].ParameterName = "@treasure";
-            myParams[2].Value = CharacterTreasure;
-
-            myParams[3] = new MySqlParameter();
-            myParams[3].ParameterName = "mid";
-            myParams[3].Value = CharacterMember;
-
-            myParams[4] = new MySqlParameter();
-            myParams[4].ParameterName = "@id";
-            myParams[4].Value = id;
-
-            db.Database.ExecuteSqlCommand(query, myParams);
-
-            return RedirectToAction("Show/" + id);
-
+            else
+            {
+                TempData["EditFail"] = "Failed to edit character";
+                return RedirectToAction("Show/" + id);
+            }
         }
 
         [Authorize(Roles = "Admin")]
         public ActionResult Delete(int id, int mid)
         {
-            string query = "delete from characters where characterid = @id";
-            db.Database.ExecuteSqlCommand(query, new MySqlParameter("@id", id));
+            if(ModelState.IsValid)
+            {
+                string query = "delete from characters where characterid = @id";
+                db.Database.ExecuteSqlCommand(query, new MySqlParameter("@id", id));
 
-            //How to redirect to another controller referenced from the following link
-            //https://stackoverflow.com/questions/10785245/redirect-to-action-in-another-controller
-            return RedirectToAction("Show/" + mid, "Member");
+                //How to redirect to another controller referenced from the following link
+                //https://stackoverflow.com/questions/10785245/redirect-to-action-in-another-controller
+                TempData["DeleteSuccess"] = "Character successfully deleted";
+                return RedirectToAction("Show/" + mid, "Member");
+
+            }
+            else
+            {
+                TempData["DeleteFail"] = "Failed to delete character";
+                return RedirectToAction("Show/" + mid, "Member");
+            }
         }
 
     }
